feat: datavolume 邏輯層調整
