+++ internal/data/kube/core.go
@@ -190,6 +190,16 @@ func (r *core) CreatePersistentVolumeClaims(ctx context.Context, config *rest.Co
 	return clientset.CoreV1().PersistentVolumeClaims(namespace).Create(ctx, pvc, opts)
 }
 
+func (r *core) GetPersistentVolumeClaims(ctx context.Context, config *rest.Config, namespace string, name string) (*oscore.PersistentVolumeClaim, error) {
+	clientset, err := r.kube.clientset(config)
+	if err != nil {
+		return nil, err
+	}
+
+	opts := metav1.GetOptions{}
+	return clientset.CoreV1().PersistentVolumeClaims(namespace).Get(ctx, name, opts)
+}
+
 func (r *core) ListPersistentVolumeClaims(ctx context.Context, config *rest.Config, namespace string) ([]oscore.PersistentVolumeClaim, error) {
 	clientset, err := r.kube.clientset(config)
 	if err != nil {
